# docs/swagger.yaml
basePath: /
definitions:
...
  admin.createAgesRequest:
    properties:
      poster:
        $ref: '#/definitions/multipart.FileHeader'
      title:
        type: string
    type: object
  admin.createCategoryRequest:
    properties:
      title:
        type: string
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
  models.Ages:
    properties:
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
    type: object
  models.ApiError:
    properties:
      error:
        type: string
    type: object
  models.Category:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
...
  multipart.FileHeader:
    properties:
      filename:
        type: string
      header:
        $ref: '#/definitions/textproto.MIMEHeader'
      size:
        type: integer
    type: object
...
  textproto.MIMEHeader:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
host: localhost:8081
info:
  contact:
    email: bexultanovmadiyar@gmail.com
    name: Madiyar Bexultanov
    url: https://www.linkedin.com/in/madiyar-bexultanov-b21902258/
  description: Personal online platform providing information about media content
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ozinshe Production
  version: "1.0"
paths:
  /admin/ages:
    get:
      description: Retrieves a list of all ages
      produces:
      - application/json
      responses:
        "200":
          description: A list of ages
          schema:
            items:
              $ref: '#/definitions/models.Ages'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all ages
      tags:
      - ages
    post:
      consumes:
      - application/json
      description: Creates a new age entry
      parameters:
      - description: Create age request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.createAgesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ID of the created age
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Failed to save poster or create age
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create new age
      tags:
      - ages
  /admin/ages/{id}:
    delete:
      description: Deletes an age by its ID
      parameters:
      - description: Age ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Success message
          schema:
            type: string
        "400":
          description: Invalid age ID
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Failed to delete age
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete an age by ID
      tags:
      - ages
    get:
      description: Retrieves an age by its ID
      parameters:
      - description: Age ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Age details
          schema:
            $ref: '#/definitions/models.Ages'
        "400":
          description: Invalid age ID
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Age not found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get age by ID
      tags:
      - ages
    put:
      consumes:
      - application/json
      description: Updates an existing age by ID
      parameters:
      - description: Age ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update age request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Ages'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            type: string
        "400":
          description: Invalid age ID or input
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Age not found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update an existing age
      tags:
      - ages
  /admin/categories:
    post:
      consumes:
      - application/json
      description: Creates a new category entry
      parameters:
      - description: Create category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.createCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ID of the created category
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Failed to create category
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create new category
      tags:
      - categories
  /admin/categories/{id}:
    delete:
      description: Deletes a category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Success message
          schema:
            type: string
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Failed to delete category
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete a category by ID
      tags:
      - categories
    get:
      description: Retrieves a category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category details
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Updates an existing category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            type: string
        "400":
          description: Invalid category ID or input
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update an existing category
      tags:
      - categories
...
  /categories:
    get:
      description: Retrieves a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: A list of categories
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all categories
      tags:
      - categories
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"